requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9

dash==2.15.0
dash-mantine-components==0.12.0
dash-draggable==0.1.2
dash-table==5.0.0
plotly==5.18.0
pandas==2.2.0

requests==2.31.0